refactor: Move orchestration methods into ReaderWindowCoordinator (#288)
