+++ minimark/Views/Window/Flow/ReaderWindowRootView+OpenAndWatchFlow.swift
@@ -1,87 +1,3 @@
+// Methods moved to ReaderWindowCoordinator (Task 6, #288).
+// This file will be deleted in Task 7.
 import Foundation
-
-extension ReaderWindowRootView {
-    func openIncomingURL(_ url: URL) {
-        guard ReaderWindowOpenAndWatchFlowSupport.isSupportedIncomingMarkdownFile(url) else {
-            return
-        }
-
-        fileOpenCoordinator.open(FileOpenRequest(
-            fileURLs: [url],
-            origin: .manual,
-            slotStrategy: .replaceSelectedSlot
-        ))
-        applyWindowTitlePresentation()
-    }
-
-    func openDocumentInCurrentWindow(_ fileURL: URL) {
-        fileOpenCoordinator.open(FileOpenRequest(
-            fileURLs: [fileURL],
-            origin: .manual,
-            folderWatchSession: folderWatchFlowController.sharedFolderWatchSession,
-            slotStrategy: .replaceSelectedSlot
-        ))
-        applyWindowTitlePresentation()
-    }
-
-    func applyInitialSeedIfNeeded() {
-        ReaderWindowOpenAndWatchFlowSupport.applyInitialSeedIfNeeded(
-            seed: seed,
-            openDocumentInCurrentWindow: { fileURL in
-                openDocumentInCurrentWindow(fileURL)
-            },
-            openDocumentInSelectedSlot: { fileURL, origin, folderWatchSession, initialDiffBaselineMarkdown in
-                openDocumentInSelectedSlot(
-                    at: fileURL,
-                    origin: origin,
-                    folderWatchSession: folderWatchSession,
-                    initialDiffBaselineMarkdown: initialDiffBaselineMarkdown
-                )
-            },
-            resolveRecentOpenedFileURL: { entry in
-                settingsStore.resolvedRecentManuallyOpenedFileURL(matching: entry.fileURL) ?? entry.fileURL
-            },
-            resolveRecentWatchedFolderURL: { entry in
-                settingsStore.resolvedRecentWatchedFolderURL(matching: entry.folderURL) ?? entry.folderURL
-            },
-            prepareRecentFolderWatch: { folderURL, options in
-                presentFolderWatchOptions(for: folderURL, options: options)
-            }
-        )
-    }
-
-    func openDocumentInSelectedSlot(
-        at fileURL: URL,
-        origin: ReaderOpenOrigin,
-        folderWatchSession: ReaderFolderWatchSession? = nil,
-        initialDiffBaselineMarkdown: String? = nil
-    ) {
-        let normalizedURL = ReaderFileRouting.normalizedFileURL(fileURL)
-        fileOpenCoordinator.open(FileOpenRequest(
-            fileURLs: [normalizedURL],
-            origin: origin,
-            folderWatchSession: folderWatchSession,
-            initialDiffBaselineMarkdownByURL: initialDiffBaselineMarkdown.map { [normalizedURL: $0] } ?? [:],
-            slotStrategy: .replaceSelectedSlot
-        ))
-        applyWindowTitlePresentation()
-    }
-
-    func prepareFolderWatchOptions(for folderURL: URL) {
-        folderWatchFlowController.prepareOptions(for: folderURL)
-    }
-
-    func presentFolderWatchOptions(for folderURL: URL, options: ReaderFolderWatchOptions) {
-        folderWatchFlowController.presentOptions(for: folderURL, options: options)
-    }
-
-    func prepareRecentFolderWatch(_ entry: ReaderRecentWatchedFolder) {
-        folderWatchFlowController.prepareRecentWatch(entry, settingsStore: settingsStore)
-    }
-
-    func updatePendingFolderWatchRequest(
-        _ update: (inout FolderWatchFlowController.PendingFolderWatchRequest) -> Void
-    ) {
-        folderWatchFlowController.updatePendingRequest(update)
-    }
-}
