+++ minimark/Views/Window/Flow/ReaderWindowRootView+WarningFlow.swift
@@ -1,54 +1,3 @@
-import AppKit
+// Methods moved to ReaderWindowCoordinator (Task 6, #288).
+// This file will be deleted in Task 7.
 import Foundation
-
-extension ReaderWindowRootView {
-    func cancelFolderWatch() {
-        folderWatchFlowController.cancelPendingWatch()
-    }
-
-    func confirmFolderWatch(_ options: ReaderFolderWatchOptions) {
-        guard let folderURL = folderWatchFlowController.pendingFolderWatchRequest?.folderURL else {
-            return
-        }
-
-        startWatchingFolder(folderURL: folderURL, options: options)
-        cancelFolderWatch()
-    }
-
-    func stopFolderWatch() {
-        dismissFolderWatchAutoOpenWarning()
-        favoriteWorkspaceController.persistFinalState(to: settingsStore)
-        favoriteWorkspaceController.deactivate()
-        groupStateController.pinnedGroupIDs = []
-        groupStateController.collapsedGroupIDs = []
-        windowCoordinator.sidebarWidth = ReaderSidebarWorkspaceMetrics.sidebarIdealWidth
-        sidebarDocumentController.folderWatchCoordinator.stopFolderWatch()
-        refreshWindowPresentation()
-        cancelFolderWatch()
-    }
-
-    func handleFolderWatchAutoOpenWarningChange(_ warning: ReaderFolderWatchAutoOpenWarning?) {
-        folderWatchFlowController.handleAutoOpenWarningChange(warning) { [self] in
-            isFolderWatchWarningPresentationAllowed()
-        }
-    }
-
-    func refreshFolderWatchAutoOpenWarningPresentation() {
-        folderWatchFlowController.refreshAutoOpenWarningPresentation { [self] in
-            isFolderWatchWarningPresentationAllowed()
-        }
-    }
-
-    func dismissFolderWatchAutoOpenWarning() {
-        folderWatchFlowController.dismissAutoOpenWarning()
-    }
-
-    func openSelectedFolderWatchAutoOpenFiles() {
-        folderWatchFlowController.openSelectedAutoOpenFiles(using: fileOpenCoordinator)
-        refreshWindowPresentation()
-    }
-
-    func isFolderWatchWarningPresentationAllowed() -> Bool {
-        folderWatchFlowController.isWarningPresentationAllowed(hostWindow: windowCoordinator.hostWindow)
-    }
-}
